PD#139454: nand: mtd&ubi support     1. enable mtd by default on p401.     2. store erase boot/data supported.     3. dtb is stored in reserved blk.     4. env is stored in reserved blk.     5. store init with erase support.     6. device_boot_flag set as MTD.     7. set mtdparts by default.     8. ubi part $partName.     9. protection on reserved blk area.     10. amlnf dtb&key erase support.     11. multi ubi device supported.     12. partition table in board config.     13. MX30LF4G18AC/MX30LF2G18AC support.     14. direct operation in amlnf cmd.

+++ fs/ubifs/ubifs.c
@@ -820,7 +820,7 @@ int ubifs_load(char *filename, u32 addr, u32 size)
 	printf("Loading file '%s' to addr 0x%08x with size %d (0x%08x)...\n",
 	       filename, addr, size, size);
 
-	page.addr = (void *)addr;
+	page.addr = (void *)(unsigned long)addr;
 	page.index = 0;
 	page.inode = inode;
 	for (i = 0; i < count; i++) {
